refactor: move components under parent Category folder
