+++ plugins/announcements/src/components/Category/CategoriesPage/CategoriesPage.tsx
@@ -11,7 +11,7 @@ import {
 import { useApi } from '@backstage/core-plugin-api';
 import { Button, makeStyles } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
-import { announcementsApiRef } from '../../api';
+import { announcementsApiRef } from '../../../api';
 import { NewCategoryDialog } from '../NewCategoryDialog';
 import { Category } from '@procore-oss/backstage-plugin-announcements-common';
