+++ plugins/announcements/src/components/Category/index.ts
@@ -1,0 +1,2 @@
+export { CategoriesPage } from './CategoriesPage';
+export { NewCategoryDialog } from './NewCategoryDialog';
